chore: move `readPackage` to `utils`
